A working first version
Resolved path planning bugs. Tested with several maze sizes and obstacle
configurations.

The algorithm is considerably slow for sizes 14x14 and bigger sizes
because of exponentially growing amount of calculations. This is a major
weakness of the original algorithm and can be improved significantly.

Also, few testcases show that the algorithm falls for local maximums
which is not desired behavior for "complex" obstacle configurations. For
random single obstacles, it most likely does not degrade performance.